+++ src/main/java/com/ddudu/application/port/in/ddudu/GetDailyDdudusByGoalUseCase.java
@@ -1,6 +1,6 @@
 package com.ddudu.application.port.in.ddudu;
 
-import com.ddudu.application.domain.ddudu.dto.response.GoalGroupedDdudus;
+import com.ddudu.application.domain.ddudu.dto.GoalGroupedDdudus;
 import java.time.LocalDate;
 import java.util.List;
